+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Fix SDL_Lock/UnlockAudio that can be called multiple times (#385)
 * Remove keyboard_support setting (#386)
 * Correctly handle opening of /dev/input/event|jsdev with writing flag
+* Fix pitch for OpenAL
 
 ## [1.4.1] - 2021-01-02
 ### Added